Fixed userDB issues

+++ app/database/userDB.tsx
@@ -6,7 +6,7 @@ type UserType = {
     password: string;
 };
 
-const openDatabase = async (): Promise<SQLite.SQLiteDatabase | null> => {
+export const openDatabase = async (): Promise<SQLite.SQLiteDatabase | null> => {
     try {
         const database = await SQLite.openDatabaseAsync('CreatureDatabase');
         console.log('Database opened successfully');
@@ -17,43 +17,73 @@ const openDatabase = async (): Promise<SQLite.SQLiteDatabase | null> => {
     }
 };
 
-const openUserTable = async (database: SQLite.SQLiteDatabase) => {
+export const openUserTable = async (databasePromise: Promise<SQLite.SQLiteDatabase | null>) => {
     try {
-        database.runAsync(`PRAGMA journal_mode = WAL`);
-        database.runAsync(`CREATE TABLE IF NOT EXISTS user (
+        const database = await databasePromise;
+
+        if (database) {
+            database.runAsync(`PRAGMA journal_mode = WAL`);
+            database.runAsync(`CREATE TABLE IF NOT EXISTS user (
           id INTEGER PRIMARY KEY NOT NULL, 
           username TEXT NOT NULL, 
           password TEXT NOT NULL
       );`);
-        console.log('Table opened successfully');
+            console.log('Table opened successfully');
+        } else {
+            console.error('Database is null, table not opened');
+        }
+
+
     } catch (error) {
         console.error('Error in openTable function', error);
     }
 };
 
-const insertUserData = async (database: SQLite.SQLiteDatabase, name: string, pass: string) => {
-    const statement = await database.prepareAsync(
-      'INSERT INTO user (username, password) VALUES ($name, $pass)'
-    );
-    try {
-      let result = await statement.executeAsync({ $name: name, $pass: pass });
-      console.log("name: " + name + " | password: " + pass, result.lastInsertRowId, result.changes);
+export const insertUserData = async (databasePromise: Promise<SQLite.SQLiteDatabase | null>, name: string, pass: string) => {
+    const database = await databasePromise;
+
+    if (database) {
+        const statement = await database.prepareAsync(
+            'INSERT INTO user (username, password) VALUES ($name, $pass)'
+        );
+
+        try {
+            let result = await statement.executeAsync({ $name: name, $pass: pass });
+            console.log("name: " + name + " | password: " + pass, result.lastInsertRowId, result.changes);
     
-    } catch (error) {
-      console.error('Error inserting data', error);
-    } finally {
-      await statement.finalizeAsync();
-      console.log('data inserted good');
-    } 
-  };
+        } catch (error) {
+            console.error('Error inserting data', error);
+        } finally {
+            await statement.finalizeAsync();
+            console.log('data inserted good');
+        }
+    } else {
+        console.error('Database is null, table not opened');
+    }
+    
+    
+};
 
-  const deleteUserData = async (database: SQLite.SQLiteDatabase, userid: number) => {
-    await database.runAsync('DELETE FROM user WHERE id = $value', { $value: userid });
-    console.log("delete good");
-  };
+export const deleteUserData = async (databasePromise: Promise<SQLite.SQLiteDatabase | null>, userid: number) => {
+    const database = await databasePromise;
 
-  const logData = async (database: SQLite.SQLiteDatabase) => {
-    const firstRow = await database.getFirstAsync('SELECT * FROM user') as UserType; 
-    console.log(firstRow.id, firstRow.username, firstRow.password);
-    console.log("good data")
-  };
+    if (database) {
+        await database.runAsync('DELETE FROM user WHERE id = $value', { $value: userid });
+        console.log("delete good");
+    } else {
+        console.error('Database is null, table not opened');
+    }
+    
+};
+
+export const logUserData = async (databasePromise: Promise<SQLite.SQLiteDatabase | null>) => {
+    const database = await databasePromise;
+
+    if (database) {
+        const firstRow = await database.getFirstAsync('SELECT * FROM user') as UserType;
+        console.log(firstRow.id, firstRow.username, firstRow.password);
+        console.log("good data");
+    } else {
+        console.error('Database is null, table not opened');
+    }
+};
